Refactor game event modules and add regression tests
